move junctions utils into top-level utils package
