+++ src/main/scala/groundtest/Tile.scala
@@ -3,9 +3,10 @@ package groundtest
 import Chisel._
 import rocket._
 import uncore.tilelink._
-import junctions._
 import scala.util.Random
 import scala.collection.mutable.ListBuffer
+import junctions.HasAddrMapParameters
+import util.ParameterizedBundle
 import cde.{Parameters, Field}
 
 case object BuildGroundTest extends Field[Parameters => GroundTest]
