+++ src/main/scala/rocket/rocket.scala
@@ -3,10 +3,11 @@
 package rocket
 
 import Chisel._
-import junctions._
 import uncore.devices._
 import uncore.agents.CacheName
 import uncore.constants._
+import junctions.HasAddrMapParameters
+import util.ParameterizedBundle
 import Util._
 import cde.{Parameters, Field}
